Added ability to change gradient orientation, background removal, default setting button and localstorage implementation
I added the ability to change the gradient orientation and the ability to remove the background image. This allows for more customization options for the user. I have also added a default setting button which allows the user to quickly reset the list to its initial state. Lastly, I have implemented localstorage to save the user's settings so that any changes made remain the same even after the page is reloaded. #gradient #bgremove #localstorage