Add ability to edit list title & Fix issue with saving edited items
This update adds the ability for the user to edit the title of the list. When the title is edited and the user confirms the changes, the new title is saved to local storage. Additionally, an issue with edited items not being saved properly has been fixed. This allows the user to have their title changes persist even after they leave the application.